Add normalize method and fix java bug
The normalize method returns a number between 0 and 1 based on the results
of the existing distance method. There are methods for Java, C and pure Ruby.

In order to implement this, we had to fix a bug in the Java code that caused
the `for` loops to skip the final character in the strings.