Change output to double rather than float
A float is good, but its precision is limited. Because of the bit
limitations of a float, the outcomes were not coming out very accurate.

This changes from using a float to a double, which increases the limit
and makes for a more accurate number when doing a normalized distance.